feat(navigation): add google maps and initial map screen
introduce a map screen with google maps with markers for both
origin and destination.

display travel path from origin to destination within the map.

add a restriction in the home screen which prevents the user from
tapping in "get a ride" or "order food" if no origin has been provided.